Site review and handlers and mailers

+++ app/views/users/partials/follow-form.blade.php
@@ -1,18 +1,20 @@
-@if($user->isFollowedBy($currentUser))
-    <p>You are following {{ $user->username }}</p>
-      {{ Form::open(['method' => 'DELETE', 'route' => ['follow_path', $user->id]]) }}
+@if($signedIn)
+    @if($user->isFollowedBy($currentUser))
+        <p>You are following {{ $user->username }}</p>
+          {{ Form::open(['method' => 'DELETE', 'route' => ['follow_path', $user->id]]) }}
+            
+                {{ Form::hidden('userIdToUnfollow', $user->id) }}
+                
+                <button type="submit" class="btn btn-danger">Unfollow {{ $user->username }}</button>
+                
+            {{ Form::close() }}
+    @else
+        {{ Form::open(['route' => 'follows_path']) }}
         
-            {{ Form::hidden('userIdToUnfollow', $user->id) }}
+            {{ Form::hidden('userIdToFollow', $user->id) }}
             
-            <button type="submit" class="btn btn-danger">Unfollow {{ $user->username }}</button>
-            
+            <button type="submit" class="btn btn-primary">Follow {{ $user->username }}</button>
+        
         {{ Form::close() }}
-@else
-    {{ Form::open(['route' => 'follows_path']) }}
-    
-        {{ Form::hidden('userIdToFollow', $user->id) }}
-        
-        <button type="submit" class="btn btn-primary">Follow {{ $user->username }}</button>
-    
-    {{ Form::close() }}
+    @endif
 @endif
